Split build target into smaller, reusable ones
Doing so means we now can reuse the logic that fixes the include::
directives in the cpython documentation in our CI job, avoiding
repeating ourselves, making maintenance easier and avoiding subtle bugs.

Being able to reuse the rest of the steps (do_build, pospell, etc)
requires more care, as the Makefile is constrained to requiring a
virtual environment (which we don't create on CI), and also doesn't
allow us to pass command-line arguments to list specific files to check.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>